Separated component into client module to add server re-direct
